download/calib/extract steps added. Additional placeholders for other steps.
