+++ src/spherical/pipeline/steps/flux_psf_calibration.py
@@ -1,0 +1,6 @@
+# Step: Flux PSF Calibration
+# This module will contain the calibrate_flux_psf logic refactored from ifs_reduction.py.
+
+def calibrate_flux_psf(converted_dir, overwrite_preprocessing, reduction_parameters):
+    # Implementation to be moved from ifs_reduction.py
+    pass
