+++ src/spherical/pipeline/steps/frame_info.py
@@ -1,0 +1,6 @@
+# Step: Frame Info Computation
+# This module will contain the frame info computation logic refactored from ifs_reduction.py.
+
+def compute_frames_info(observation, converted_dir):
+    # Implementation to be moved from ifs_reduction.py
+    pass
