+++ src/spherical/pipeline/steps/plot_center_evolution.py
@@ -1,0 +1,6 @@
+# Step: Plot Image Center Evolution
+# This module will contain the plot_image_center_evolution logic refactored from ifs_reduction.py.
+
+def plot_image_center_evolution(converted_dir):
+    # Implementation to be moved from ifs_reduction.py
+    pass
