+++ src/spherical/pipeline/steps/process_centers.py
@@ -1,0 +1,6 @@
+# Step: Process Extracted Centers
+# This module will contain the process_extracted_centers logic refactored from ifs_reduction.py.
+
+def process_extracted_centers(converted_dir, extraction_parameters, non_least_square_methods, overwrite_preprocessing):
+    # Implementation to be moved from ifs_reduction.py
+    pass
