+++ src/spherical/pipeline/steps/spot_photometry.py
@@ -1,0 +1,6 @@
+# Step: Spot Photometry Calibration
+# This module will contain the calibrate_spot_photometry logic refactored from ifs_reduction.py.
+
+def calibrate_spot_photometry(converted_dir, overwrite_preprocessing):
+    # Implementation to be moved from ifs_reduction.py
+    pass
