+++ src/spherical/pipeline/steps/spot_to_flux.py
@@ -1,0 +1,6 @@
+# Step: Spot-to-Flux Normalization
+# This module will contain the spot_to_flux logic refactored from ifs_reduction.py.
+
+def spot_to_flux_normalization(converted_dir, reduction_parameters):
+    # Implementation to be moved from ifs_reduction.py
+    pass
